Clean up load/weather additions to satisfy pylint and conventions
Follow-up cleanup on top of the Load (#8) and Weather (#9) contributions so
the package passes the pylint CI gate (fail-under=10) and matches existing
conventions:

- Extract a shared VipResource base for the identical get_voltage/current/
  power accessors on Grid and Load (resolves the duplicate-code failure).
- Move the datetime import in client.py to module top-level
  (import-outside-toplevel).
- Strip trailing whitespace and trailing newline in plant.py.
- Tidy stray blank lines in client.py and the test files.
- Add a Weather empty-data test (brings weather.py to 100% coverage).
- Document get_inverter_realtime_load, get_plant and get_weather in the README.
